requirements.txt
pdfplumber==0.11.9
sentence-transformers
faiss-cpu
transformers
torch
